Reformatting package structure
