--- app/src/main/java/com/dlight/algoguide/algorithms/searching/linear_search/LinearSearch.kt
@@ -1,14 +1,0 @@
-package com.dlight.algoguide.algorithms.searching.linear_search
-
-class LinearSearch {
-     suspend fun linearSearch(arr: IntArray, target: Int, isFound: (Boolean) -> Unit){
-        for(i in arr){
-            if(i == target) {
-                isFound(true)
-                break
-            }
-            isFound(false)
-        }
-    }
-}
-
