--- app/src/main/java/com/dlight/algoguide/algorithms/sorting/Events.kt
@@ -1,9 +1,0 @@
-package com.dlight.algoguide.algorithms.sorting
-
-sealed class Events {
-    object speedUp : Events()
-    object slowDown : Events()
-    object playPause : Events()
-    object previous : Events()
-    object next : Events()
-}
